moved object class files to domain package
